Revised the cp2k pdos calculations, added the spectrum calculation module - for cp2k, fixed some minor bug in normal modes scripts

+++ src/libra_py/normal_modes.py
@@ -123,7 +123,7 @@ def visualize_modes(E, R, U, M, w, params):
     """
 
     ndof = R.num_of_rows
-    nat = ndof/3
+    nat = int(ndof/3)
 
     scl = params["scale"]
